+++ scripts/__pycache__/buddha.cpython-39.pyc
