+++ scripts/__pycache__/ganesh.cpython-38.pyc
