+++ scripts/__pycache__/ganesh.cpython-39.pyc
