+++ scripts/ganesh.py
